Login/requirements.txt
flask
mysql-connector
docker-compose
flask-cors